Added the assert property ("power") for the MotorSubsystem.
Added asserts in the simulation files you had created to check motor states.  See the
simulation files for examples of how this works.  We need to add the capability of instead
of supplying a value, comparing to a value in the settings file.  The way it works now if
you update the intake power from 1.0 to 0.9 for instance, we will have to change all of the
simulation files that check for 1.0 as the power of the intake.

Shortened the simulation times to better match each test.  When you run these over and over
you want the times to be as short as possible without being too short.  We need to add the
error Mark mentioned yesterday, that if the simulation time ends with events left in the
simulation file, it fails and prints an error.

Copied the test runner scripts from droid2021 to the bunny bot top level directory
to run suites of tests.
  runall - will build and run all tests (i don't use this very much)
  runtests - will NOT build but will run all tests
             or if you add the --list command line argument will list the available tests
             or if you add a keyword it will run tests that match the keyword
Note, when you run using the scripts, the output will be in a directory named testout.  There
will be two files named after the test.  For instance, if the test is 1_teleop_intake, then
there will be 1_teleop_intake.out and 1_teleop_intake.log.  The log file is obvious.  The out
file is what the robot would have printed to standard "output".

Renamed the simulation files to include "teleop" in the name so that when using test runner
I could do "./runtests teleop" to get all of these teleop tests.